requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.11.18 ; python_version >= "3.11" and python_version < "4.0"
aioice==0.10.1 ; python_version >= "3.11" and python_version < "4.0"
aiortc==1.12.0 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.11" and python_version < "4.0"
av==14.3.0 ; python_version >= "3.11" and python_version < "4.0"
azure-core==1.34.0 ; python_version >= "3.11" and python_version < "4.0"
azure-identity==1.22.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.4.26 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0"
chardet==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.11" and python_version < "4.0"
cloudevents==1.11.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
cryptography==44.0.3 ; python_version >= "3.11" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
deprecated==1.2.18 ; python_version >= "3.11" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.2.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
ifaddr==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.11" and python_version < "4.0"
isodate==0.7.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.9.0 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-path==0.3.4 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2025.4.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.11" and python_version < "4.0"
lazy-object-proxy==1.11.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
more-itertools==10.7.0 ; python_version >= "3.11" and python_version < "4.0"
msal-extensions==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
msal==1.32.3 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.4.3 ; python_version >= "3.11" and python_version < "4.0"
nest-asyncio==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.2.5 ; python_version >= "3.11" and python_version < "4.0"
openai==1.78.1 ; python_version >= "3.11" and python_version < "4.0"
openapi-core==0.19.5 ; python_version >= "3.11" and python_version < "4.0"
openapi-schema-validator==0.6.3 ; python_version >= "3.11" and python_version < "4.0"
openapi-spec-validator==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-api==1.33.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-sdk==1.33.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-semantic-conventions==0.54b0 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
parse==1.20.2 ; python_version >= "3.11" and python_version < "4.0"
pathable==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
prance==25.4.8.0 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
psycopg-binary==3.2.7 ; python_version >= "3.11" and python_version < "4.0"
psycopg==3.2.7 ; python_version >= "3.11" and python_version < "4.0"
pybars4==0.9.13 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.9.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.4 ; python_version >= "3.11" and python_version < "4.0"
pyee==13.0.0 ; python_version >= "3.11" and python_version < "4.0"
pyjwt[crypto]==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pylibsrtp==0.12.0 ; python_version >= "3.11" and python_version < "4.0"
pymeta3==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
pyopenssl==25.0.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.36.2 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rfc3339-validator==0.1.4 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.24.0 ; python_version >= "3.11" and python_version < "4.0"
ruamel-yaml-clib==0.2.12 ; platform_python_implementation == "CPython" and python_version < "3.13" and python_version >= "3.11"
ruamel-yaml==0.18.10 ; python_version >= "3.11" and python_version < "4.0"
scipy==1.15.3 ; python_version >= "3.11" and python_version < "4.0"
semantic-kernel==1.29.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.40 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
wrapt==1.17.2 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.20.0 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.11" and python_version < "4.0"
# Web Application Dependencies
streamlit>=1.28.0
pandas>=2.0.0
polars>=0.20.0

# Azure Service Bus
azure-servicebus>=7.11.0

# Environment Management
python-dotenv>=1.0.0

# Additional utilities for the webapp
plotly>=5.17.0  # For potential visualizations
requests>=2.31.0  # For HTTP requests if needed
